Move interface to a new file that is ignored by coverage Thanks @hynek
