+++ tox.ini
@@ -18,7 +18,7 @@ deps =
 passenv = TERM  # ensure colors
 commands =
     coverage run --branch --source=treq {envbindir}/trial treq
-    coverage report -m
+    coverage report --omit=treq/interfaces.py -m
 
 [testenv:pypi-readme]
 deps =
